chore: add moti-sync.sh script for local development from ./examples/sample
Run this script to watch the `packages/moti/src` folder for changes and automatically sync those changes into the `node_modules` of the sample project.

I also updated the sample project to stop using `packages/moti` directly, since it was leading to module resolution issues. Therefore, the only way to test the local `moti` is to run `yarn start` in one terminal tab and `yarn moti:sync` in another.